Abstract add_signal_handler utility
Specifically from the dsim, fgpu and xbgpu into a central utils.py.

Resolves: NGC-534.
