+++ src/katgpucbf/utils.py
@@ -1,0 +1,42 @@
+################################################################################
+# Copyright (c) 2020-2022, National Research Foundation (SARAO)
+#
+# Licensed under the BSD 3-Clause License (the "License"); you may not use
+# this file except in compliance with the License. You may obtain a copy
+# of the License at
+#
+#   https://opensource.org/licenses/BSD-3-Clause
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+################################################################################
+
+"""A collection of utility functions for katgpucbf."""
+
+import logging
+import signal
+from asyncio import get_event_loop
+
+from aiokatcp import DeviceServer
+
+logger = logging.getLogger(__name__)
+
+
+def add_signal_handlers(server: DeviceServer) -> None:
+    """Arrange for clean shutdown on SIGINT (Ctrl-C) or SIGTERM."""
+    signums = [signal.SIGINT, signal.SIGTERM]
+
+    def handler():
+        # Remove the handlers so that if it fails to shut down, the next
+        # attempt will try harder.
+        logger.info("Received signal, shutting down")
+        for signum in signums:
+            loop.remove_signal_handler(signum)
+        server.halt()
+
+    loop = get_event_loop()
+    for signum in signums:
+        loop.add_signal_handler(signum, handler)
